# packages/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: adres-pl — mikroserwis adresowy
  version: '2.0'
  description: |
    Podpowiedzi adresowe, walidacja i geokodowanie odwrotne na danych
    Państwowego Rejestru Granic (PRG) i TERYT.

    **Uwierzytelnianie: klucz API w nagłówku `X-API-Key`.** Klucz ma postać
    `adr_live_<32 znaki>_<6 znaków sumy>` i jest pokazywany **jeden raz**,
    przy wystawieniu — serwis przechowuje wyłącznie jego skrót.

    Limit zapytań liczony jest po **zweryfikowanym kliencie**, nie po kluczu
    ani po nagłówku: wszystkie klucze jednego klienta dzielą wspólny licznik.
    Ruch bez ważnego klucza podlega osobnemu, niższemu limitowi po adresie
    sieciowym. Limit obowiązuje **na instancję usługi** — przy wielu replikach
    łączna przepustowość jest odpowiednio wyższa.

    Sondy operacyjne (`/health`, `/ready`, `/metrics`, `/status`) pozostają
    bez uwierzytelniania: są potrzebne Kubernetesowi i Prometheuszowi, a odcina
    się je na warstwie sieciowej, nie w aplikacji.

    **Wersja danych** jest niezależna od wersji API. Zwracają ją `/v1/meta`
    oraz pola `wersjaDanych` w odpowiedziach wyszukiwania i walidacji.
  license:
    name: Dane PRG bez licencji (art. 40c ust. 5 Prawa geodezyjnego)
servers:
  - url: http://localhost:3000
    description: Instalacja lokalna

# Obowiazuje wszystkie operacje POZA tymi, ktore maja jawne `security: []`
# (sondy operacyjne). Wpisanie tego globalnie, a nie przy kazdej operacji,
...
# napisac `security: []`, zeby go otworzyc.
security:
  - ApiKey: []
...
paths:
...
  /admin/clients:
    post:
      tags: [administrative]
      summary: Zakłada klienta
      security: [{ TokenOperatora: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                nip: { type: string, pattern: '^[0-9]{10}$' }
                email: { type: string }
                plan: { type: string, default: test }
                rateLimitPerMin: { type: integer, default: 600 }
                monthlyQuota: { type: integer, description: W jednostkach rozliczeniowych. }
      responses:
        '200': { description: Klient założony. }
        '201': { description: Klient założony. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
    get:
      tags: [administrative]
      summary: Lista klientów
      security: [{ TokenOperatora: [] }]
      responses:
        '200': { description: Klienci wraz z limitami i kwotami. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
  /admin/keys:
    post:
      tags: [administrative]
      summary: Wystawia klucz — jedyne miejsce zwracające klucz jawny
      description: |
        Klucz jawny jest zwracany **raz** i nie da się go odtworzyć — w bazie
        leży wyłącznie skrót. Odpowiedź ma `Cache-Control: no-store`, bo jej
        treść jest poświadczeniem.
      security: [{ TokenOperatora: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId: { type: integer }
                environment: { type: string, enum: [live, test], default: live }
                name: { type: string }
                replacesId: { type: integer }
      responses:
        '200': { description: Klucz wystawiony. }
        '201': { description: Klucz wystawiony. Pole `klucz` widoczne jedyny raz. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
    get:
      tags: [administrative]
      summary: Lista kluczy — nigdy nie zwraca klucza ani skrótu
      security: [{ TokenOperatora: [] }]
      responses:
        '200': { description: Klucze wraz ze stanem i zużyciem w bieżącym okresie. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
  /admin/keys/rotate:
    post:
      tags: [administrative]
      summary: Rotacja bezprzerwowa — następca obok, poprzednik z terminem
      description: |
        Poprzednik dostaje **termin**, nie natychmiastowe unieważnienie: przez
        okres przejściowy działają oba klucze. Powiadomienie klienta jest
        warunkiem przejścia dalej — patrz `docs/runbook-klucze.md`.
      security: [{ TokenOperatora: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKeyId]
              properties:
                apiKeyId: { type: integer }
                graceDays: { type: integer, default: 7 }
      responses:
        '200': { description: Następca wystawiony. }
        '201': { description: Następca wystawiony. Pole `klucz` widoczne jedyny raz. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
        '404': { $ref: '#/components/responses/BadRequest' }
  /admin/keys/revoke:
    post:
      tags: [administrative]
      summary: Unieważnia klucz natychmiast
      description: Zbieżność we wszystkich instancjach typowo poniżej 100 ms.
      security: [{ TokenOperatora: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKeyId]
              properties:
                apiKeyId: { type: integer }
                reason: { type: string }
      responses:
        '200': { description: Klucz unieważniony. }
        '401': { $ref: '#/components/responses/MissingOperatorToken' }
        '404': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    OperatorToken:
      type: http
      scheme: bearer
      description: |
        Token operatora do zarządzania klientami i kluczami. **Odrębny od kluczy
        klienckich**: klucz `adr_live_*` nie otwiera żadnej trasy `/admin` —
        inaczej dowolny klient wystawiłby sobie klucz bez limitów.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Klucz klienta. **Wyłącznie w nagłówku** — klucz podany w query stringu
        jest traktowany jak jego brak, bo trafiłby do access logu ingressu,
        do nagłówka `Referer` i do historii przeglądarki.
  responses:
    MissingOperatorToken:
      description: 'Brak nagłówka `Authorization: Bearer` albo token niezgodny.'
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string, example: 'Wymagany token operatora.' }
              code: { type: string, enum: [BRAK_TOKENU] }
...
    BadRequest:
      description: Żądanie poprawne składniowo, ale niewykonalne.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string, example: 'Podaj ulicId albo simc.' }
...